# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LoanType = typeof LoanType[keyof typeof LoanType];


export const LoanType = {
  Personal: 'Personal',
  Car: 'Car',
  Housing: 'Housing',
  Education: 'Education',
} as const;

export interface Bank {
  id: string;
  name: string;
  /** @nullable */
  logo_url: string | null;
}

export interface LoanProduct {
  id: string;
  bank_id: string;
  loan_type: LoanType;
  /** @minimum 0 */
  min_interest_rate: number;
  /** @minimum 0 */
  max_interest_rate: number;
  /** @minimum 1 */
  max_tenure_months: number;
  /** @minimum 1 */
  max_loan_amount: number;
  /** @nullable */
  processing_fee_desc: string | null;
  features: string[];
  bank: Bank;
}

export interface LoanApplicationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  user_name: string;
  /** @maxLength 320 */
  email: string;
  /** @pattern ^\d{10}$ */
  phone: string;
  loan_product_id: string;
  /** @exclusiveMinimum 0 */
  requested_amount: number;
  /** @minimum 1 */
  tenure_months: number;
}

export interface LoanApplicationReceipt {
  id: string;
  /** @minimum 0 */
  estimated_emi: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListLoanProductsParams = {
loanType: LoanType;
};