# lib/api-zod/src/generated/types/bank.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Bank {
  id: string;
  name: string;
  /** @nullable */
  logo_url: string | null;
}