# lib/api-zod/src/generated/types/loanApplicationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoanApplicationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  user_name: string;
  /** @maxLength 320 */
  email: string;
  /** @pattern ^\d{10}$ */
  phone: string;
  loan_product_id: string;
  /** @exclusiveMinimum 0 */
  requested_amount: number;
  /** @minimum 1 */
  tenure_months: number;
}