# lib/api-zod/src/generated/types/loanApplicationReceipt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoanApplicationReceipt {
  id: string;
  /** @minimum 0 */
  estimated_emi: number;
}