# lib/api-zod/src/generated/types/loanProduct.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Bank } from './bank';
import type { LoanType } from './loanType';

export interface LoanProduct {
  id: string;
  bank_id: string;
  loan_type: LoanType;
  /** @minimum 0 */
  min_interest_rate: number;
  /** @minimum 0 */
  max_interest_rate: number;
  /** @minimum 1 */
  max_tenure_months: number;
  /** @minimum 1 */
  max_loan_amount: number;
  /** @nullable */
  processing_fee_desc: string | null;
  features: string[];
  bank: Bank;
}