# lib/api-zod/src/generated/types/loanType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type LoanType = typeof LoanType[keyof typeof LoanType];


export const LoanType = {
  Personal: 'Personal',
  Car: 'Car',
  Housing: 'Housing',
  Education: 'Education',
} as const;